Separated tests for __main__.py as requested
